+++ tests/scalers/redis-cluster-lists.test.ts
@@ -19,7 +19,7 @@ const redisWorkerHostPortRefDeploymentName = 'redis-worker-test-hostport'
 const redisWorkerAddressRefDeploymentName = 'redis-worker-test-address'
 const redisWorkerHostPortRefTriggerAuthDeploymentName = 'redis-worker-test-hostport-triggerauth'
 const itemsToWrite = 200
-const deploymentContainerImage = 'goku321/redis-cluster-list:v1.4'
+const deploymentContainerImage = 'goku321/redis-cluster-list:v1.7'
 const writeJobNameForHostPortRef = 'redis-writer-host-port-ref'
 const writeJobNameForAddressRef = 'redis-writer-address-ref'
 const writeJobNameForHostPortInTriggerAuth = 'redis-writer-host-port-trigger-auth'
@@ -28,9 +28,11 @@ test.before(t => {
     // Deploy Redis cluster.
     sh.exec(`kubectl create namespace ${redisNamespace}`)
     sh.exec(`helm repo add bitnami https://charts.bitnami.com/bitnami`)
+
+    let clusterStatus = sh.exec(`helm install --timeout 600s ${redisClusterName} --namespace ${redisNamespace} --set "global.redis.password=${redisPassword}" bitnami/redis-cluster`).code
     t.is(0,
-      sh.exec(`helm install ${redisClusterName} --namespace ${redisNamespace} --set "global.redis.password=${redisPassword}" bitnami/redis-cluster`).code,
-      'creating a Redis cluster should work.'
+        clusterStatus,
+        'creating a Redis cluster should work.'
     )
 
     // Wait for Redis cluster to be ready.
@@ -137,7 +139,7 @@ test.serial(`Deployment using redis host port env vars should max and scale to 5
     runWriteJob(t, writeJobNameForHostPortRef, listNameForHostPortRef)
 
     let replicaCount = '0'
-    for (let i = 0; i < 20 && replicaCount !== '5'; i++) {
+    for (let i = 0; i < 30 && replicaCount !== '5'; i++) {
         replicaCount = sh.exec(
             `kubectl get deployment/${redisWorkerHostPortRefDeploymentName} --namespace ${testNamespace} -o jsonpath="{.spec.replicas}"`
         ).stdout
@@ -177,7 +179,7 @@ test.serial(`Deployment using redis address env var should max and scale to 5 wi
     runWriteJob(t, writeJobNameForAddressRef, listNameForAddressRef)
 
     let replicaCount = '0'
-    for (let i = 0; i < 20 && replicaCount !== '5'; i++) {
+    for (let i = 0; i < 30 && replicaCount !== '5'; i++) {
         replicaCount = sh.exec(
             `kubectl get deployment/${redisWorkerAddressRefDeploymentName} --namespace ${testNamespace} -o jsonpath="{.spec.replicas}"`
         ).stdout
@@ -217,7 +219,7 @@ test.serial(`Deployment using redis host port in triggerAuth should max and scal
     runWriteJob(t, writeJobNameForHostPortInTriggerAuth, listNameForHostPortTriggerAuth)
 
     let replicaCount = '0'
-    for (let i = 0; i < 20 && replicaCount !== '5'; i++) {
+    for (let i = 0; i < 30 && replicaCount !== '5'; i++) {
         replicaCount = sh.exec(
             `kubectl get deployment/${redisWorkerHostPortRefTriggerAuthDeploymentName} --namespace ${testNamespace} -o jsonpath="{.spec.replicas}"`
         ).stdout
@@ -327,7 +329,7 @@ spec:
         - name: REDIS_PASSWORD
           value: {{REDIS_PASSWORD}}
         - name: READ_PROCESS_TIME
-          value: "200"
+          value: "500"
 ---
 apiVersion: keda.sh/v1alpha1
 kind: ScaledObject
@@ -437,7 +439,7 @@ spec:
         - name: REDIS_PASSWORD
           value: {{REDIS_PASSWORD}}
         - name: READ_PROCESS_TIME
-          value: "200"
+          value: "500"
 ---
 apiVersion: keda.sh/v1alpha1
 kind: ScaledObject